Shift the Todo Item List into a component.
