+++ Frontend-React/src/types/index.d.ts
@@ -1,0 +1,5 @@
+export interface TodoItem {
+    id: number
+    description: string
+    completed: boolean
+  }
